benchmarks/requirements.txt
aiohttp>=3.8.0
psutil>=5.9.0
requests>=2.28.0
Pillow>=9.0.0
numpy>=1.21.0